LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/elementsFromPoint-svg-text-expected.txt

Some text
Some text
Some text
Text underText over

PASS elementsFromPoint for a point inside a <text>
PASS elementsFromPoint for a point inside a <tspan> nested in a <text> without content
PASS elementsFromPoint for a point inside a <textPath> nested in a <text> without content
FAIL elementsFromPoint for a point inside an overlapping <tspan> nested in a <text> assert_equals: document.elementsFromPoint(125,305) expected "tspan#tspan2, text#text4, svg#svg, DIV#sandbox, BODY, HTML" but got "text#text4, tspan#tspan2, svg#svg, DIV#sandbox, BODY, HTML"
